requirements.txt
openai>=1.0.0
anthropic>=0.25.0
python-dotenv>=1.0.0
click>=8.0.0
colorama>=0.4.0